Move filesystem drivers into their own subdirectories
