+++ kernel/src/fs/devfs/subdir.mk
@@ -1,0 +1,2 @@
+
+local_sources := devfs.c
