+++ kernel/src/fs/ext2/subdir.mk
@@ -1,0 +1,2 @@
+
+local_sources := ext2.c
