+++ kernel/src/fs/subdir.mk
@@ -1,1 +1,3 @@
-local_sources := vfs.c ext2.c devfs.c
+
+local_sources := vfs.c
+dirs := devfs ext2
